+++ src/luxonis_ml/data/labelstudio.py
@@ -1,76 +1,11 @@
 from .dataset import LuxonisDataset
+import luxonis_ml.data.utils.labelstudio as lsUtils
 from label_studio_sdk import Client, Project
 
-# from label_studio_converter import brush
-from typing import List, Union
-from pycocotools import mask as maskUtils
+from typing import List
 import numpy as np
 import os
 import xml.etree.ElementTree as ET
-
-
-class InputStream:
-    def __init__(self, data):
-        self.data = data
-        self.i = 0
-
-    def read(self, size):
-        out = self.data[self.i : self.i + size]
-        self.i += size
-        return int(out, 2)
-
-
-def access_bit(data, num):
-    """from bytes array to bits by num position"""
-    base = int(num // 8)
-    shift = 7 - int(num % 8)
-    return (data[base] & (1 << shift)) >> shift
-
-
-def bytes2bit(data):
-    """get bit string from bytes data"""
-    return "".join([str(access_bit(data, i)) for i in range(len(data) * 8)])
-
-
-def rle_to_mask(rle: List[int], height: int, width: int) -> np.array:
-    """
-    Converts rle to image mask
-    Args:
-        rle: your long rle
-        height: original_height
-        width: original_width
-
-    Returns: np.array
-    """
-
-    rle_input = InputStream(bytes2bit(rle))
-
-    num = rle_input.read(32)
-    word_size = rle_input.read(5) + 1
-    rle_sizes = [rle_input.read(4) + 1 for _ in range(4)]
-    # print('RLE params:', num, 'values,', word_size, 'word_size,', rle_sizes, 'rle_sizes')
-
-    i = 0
-    out = np.zeros(num, dtype=np.uint8)
-    while i < num:
-        x = rle_input.read(1)
-        j = i + 1 + rle_input.read(rle_sizes[rle_input.read(2)])
-        if x:
-            val = rle_input.read(word_size)
-            out[i:j] = val
-            i = j
-        else:
-            while i < j:
-                val = rle_input.read(word_size)
-                out[i] = val
-                i += 1
-
-    image = np.reshape(out, [height, width, 4])[:, :, 3]
-    return image
-
-
-def generate_random_color():
-    return [np.random.randint(0, 255) for _ in range(3)]
 
 
 class LabelStudioConnector:
@@ -138,7 +73,7 @@ class LabelStudioConnector:
                         box_cls = ann["value"]["rectanglelabels"][0]
                         boxes.append([box_cls, x, y, w, h])
                     if ann["type"] == "brushlabels":
-                        mask = rle_to_mask(
+                        mask = lsUtils.rle_to_mask(
                             ann["value"]["rle"],
                             ann["original_height"],
                             ann["original_width"],
@@ -165,66 +100,8 @@ class LabelStudioConnector:
 
         self.dataset.add(additions, media_exists=True)
 
-    def get_xml_config(self) -> str:
-        root = ET.Element("View")
-
-        image = ET.SubElement(root, "Image")
-        image.set("name", "image")
-        image.set("value", "$image")
-        image.set("zoom", "true")
-
-        classes = self.dataset.fo_dataset.classes
-
-        if "class" in classes and len(classes["class"]):
-            choices = ET.SubElement(root, "Choices")
-            choices.set("name", "choice")
-            choices.set("toName", "image")
-            for label in classes["class"]:
-                choice = ET.SubElement(choices, "Choice")
-                choice.set("value", label)
-
-        if "boxes" in classes and len(classes["boxes"]):
-            inst = ET.SubElement(root, "RectangleLabels")
-            inst.set("name", "label")
-            inst.set("toName", "image")
-            for label in classes["boxes"]:
-                r, g, b = generate_random_color()
-                lbl = ET.SubElement(inst, "Label")
-                lbl.set("value", label)
-                lbl.set("background", f"rgba({r},{b},{g},0.7)")
-
-        if "segmentation" in classes and len(classes["segmentation"]):
-            inst = ET.SubElement(root, "BrushLabels")
-            inst.set("name", "tag")
-            inst.set("toName", "image")
-            for label in classes["segmentation"]:
-                r, g, b = generate_random_color()
-                lbl = ET.SubElement(inst, "Label")
-                lbl.set("value", label)
-                lbl.set("background", f"rgba({r},{b},{g},0.7)")
-
-        if "keypoints" in classes and len(classes["keypoints"]):
-            inst = ET.SubElement(root, "KeyPointLabels")
-            inst.set("name", "kp-1")
-            inst.set("toName", "image")
-            for label in classes["keypoints"]:
-                r, g, b = generate_random_color()
-                lbl = ET.SubElement(inst, "Label")
-                lbl.set("value", label)
-                lbl.set("background", f"rgba({r},{b},{g},0.7)")
-
-        tree = ET.ElementTree(root)
-        tree.write("tmp.xml", encoding="utf-8", xml_declaration=False)
-
-        with open("tmp.xml") as file:
-            label_config_str = file.read()
-
-        os.remove("tmp.xml")
-
-        return label_config_str
-
     def create_project(self, project_name: str):
-        label_config_str = self.get_xml_config()
+        label_config_str = lsUtils.get_xml_config()
 
         project = Project(
             self.dataset._get_credentials("LABELSTUDIO_URL"),
